rag/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7

pypdf==4.3.1
python-docx==1.1.2

faiss-cpu==1.7.4

sentence-transformers==3.0.1
torch==2.3.1
transformers==4.43.3
tokenizers==0.19.1

numpy==1.26.4